Rename GLS to GMean and update tests
