+++ CHANGELOG.md
@@ -10,7 +10,7 @@ changelog does not include internal changes that do not affect the user.
 
 ### Added
 
-- Added `GLS` studied in [MultiNet++: Multi-Stream Feature Aggregation and Geometric Loss Strategy for
+- Added `GMean` studied in [MultiNet++: Multi-Stream Feature Aggregation and Geometric Loss Strategy for
   Multi-Task Learning](https://openaccess.thecvf.com/content_CVPRW_2019/papers/WAD/Chennupati_MultiNet_Multi-Stream_Feature_Aggregation_and_Geometric_Loss_Strategy_for_Multi-Task_CVPRW_2019_paper.pdf),
   a `Scalarizer` that returns the geometric mean of the input tensor of values.
