--- docs/source/docs/scalarization/gls.rst
@@ -1,7 +1,0 @@
-:hide-toc:
-
-GLS
-===
-
-.. autoclass:: torchjd.scalarization.GLS
-    :members: __call__
