+++ docs/source/docs/scalarization/gmean.rst
@@ -1,0 +1,7 @@
+:hide-toc:
+
+GMean
+=====
+
+.. autoclass:: torchjd.scalarization.GMean
+    :members: __call__
